+++ changelog.md
@@ -5,6 +5,10 @@ next
 * Improve the error messages that are thrown when forcing nonexistent
   benchmark environments.
 
+* Explicitly mark `forceGC` as deprecated. `forceGC` has not had any effect
+  for several releases, and it will be removed in the next major `criterion`
+  release.
+
 1.2.2.0
 
 * Important bugfix: versions 1.2.0.0 and 1.2.1.0 were incorrectly displaying